Abstract/Refactor walkthrough code into separate file, add comments
