+++ frontend/src/LandingPage.tsx
@@ -98,7 +98,7 @@ function LandingPage(
     joyrideState: JoyrideState,
     setJoyrideState: React.Dispatch<React.SetStateAction<JoyrideState>>
   }) {
-    
+
     const [sheetId, setSheetId] = useState("");
     // const [sheetURL, setSheetURL] = useState("");
 
@@ -114,6 +114,8 @@ function LandingPage(
         props.setRateData(getRates());
         props.setVisibility(true);
         if (props.joyrideState.tourActive) {
+          // Progress through walkthrough after pause
+          // Note: hardcoded stepIndex based on steps in Walkthrough.tsx
           props.setJoyrideState(prevState => ({
             ...prevState,
             run: true,
@@ -129,6 +131,7 @@ function LandingPage(
         setSheetId(event.target.value.split('/')[5]);
     }
 
+    // Start walkthrough
     const helpButtonClicked = () => {
       props.setJoyrideState(prevState => ({
         ...prevState,
